scripts/multi-agent-setup/requirements.txt
# OpenAI Agents SDK and dependencies
openai-agents>=0.1.0
openai>=1.0.0

# Environment variable management
python-dotenv>=1.0.0

# Async support (usually included with Python 3.7+)
asyncio

# Optional: for better async support
aiohttp>=3.8.0